#2561 - changed getAverageMsRuntimeForQuiz to return a Double and then convert that to an Integer to prevent ConversionFailedException since the result can potentially be a Double

+++ service/src/main/java/skills/storage/repos/UserQuizAttemptRepo.groovy
@@ -51,7 +51,7 @@ interface UserQuizAttemptRepo extends JpaRepository<UserQuizAttempt, Long> {
                     and quizDef.quiz_id = ?1
                     and quizAttempt.status <> 'INPROGRESS'
      ''', nativeQuery = true)
-    Integer getAverageMsRuntimeForQuiz(String quizId)
+    Double getAverageMsRuntimeForQuiz(String quizId)
 
 
     @Query('''select count(distinct quizAttempt.userId)
